handle hashing larger files

+++ tests/dgst/dgst-test.sh
@@ -45,6 +45,21 @@ done
 run "dgst -sha256 -verify ./certs/server-keyPub.pem -signature ./tests/dgst/5000-server-key.sig ./large-test.txt"
 run "dgst -sha256 -sign ./certs/server-key.pem -out 5000-server-key.sig ./large-test.txt"
 run "dgst -sha256 -verify ./certs/server-keyPub.pem -signature ./5000-server-key.sig ./large-test.txt"
+
+# run some hash tests on large file while available
+run "-hash sha256 -in ./large-test.txt"
+echo $RESULT | grep "3e5915162b1974ac0d57a5a45113a1efcc1edc5e71e5e55ca69f9a7c60ca11fd"
+if [ $? -ne 0 ]; then
+    echo "Failed to get expected hash of large file with -hash"
+    exit 99
+fi
+run "sha256 ./large-test.txt"
+echo $RESULT | grep "3e5915162b1974ac0d57a5a45113a1efcc1edc5e71e5e55ca69f9a7c60ca11fd"
+if [ $? -ne 0 ]; then
+    echo "Failed to get expected hash of large file with sha256"
+    exit 99
+fi
+
 rm -f large-test.txt
 
 run "dgst -sha256 -sign ./certs/ecc-key.pem -out configure.sig configure.ac"
